Frontend fixes.
